Squashed commit of the following:
commit b884e3c1242c52c0209f17601de35f86a6bc3260
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Thu May 9 17:48:38 2024 +0800

    edit the tempData[Error] to Info

commit 4a4dfdef820b5cbc6f1d4ac0d951fb7e6ed41a08
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Thu May 9 17:48:11 2024 +0800

    make the temp["info"] able to work

commit b23fa11ccc51693aa25b42bfdfd42baefc7bd107
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Thu May 9 17:40:16 2024 +0800

    Subjects CRUD

commit 2e42c2435359d8b315c80352e7740b35e64a8f29
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Thu May 9 17:39:50 2024 +0800

    Add classSubjects

commit 4ac9cfe081124e8f0a28feb42234eb836d77f486
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Mon May 6 14:27:24 2024 +0800

    subject CR

    create for getRequest, Read also only getRequest

commit e2f3f0b49f5e45d80df9b1f97a7c9056ff939716
Merge: 2dfa27b 0275368
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Mon May 6 14:26:51 2024 +0800

    Merge branch 'Subject-Class-Allocation' of https://github.com/simony666/EWA into Subject-Class-Allocation

commit 2dfa27bdb06f41b9c6a8fa69eb80efca3c34f15d
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Mon May 6 14:18:24 2024 +0800

    change favicon

commit 15713c8193a7a3465e6a3599b520ffb29c76e10c
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Mon May 6 14:18:13 2024 +0800

    change URL

commit da29714fff76d81663bd25a76462c195fb0c0b8c
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Mon May 6 14:17:57 2024 +0800

    classes CRUD

commit 027536899dfc0f2b13601f37ffd901bb4a0520f8
Author: Simon Yong <51854826+simony666@users.noreply.github.com>
Date:   Mon May 6 13:58:47 2024 +0800

    Update Staff UI

commit 5770a32bc996718ff2184a1bdd3fcfbda5007356
Author: Simon Yong <51854826+simony666@users.noreply.github.com>
Date:   Mon May 6 13:58:20 2024 +0800

    Delete File and update git ignore

commit f31d93e0872d245009272ea90a77bf68120da4cd
Author: Zzz-anonymous <firsttry2468@gmail.com>
Date:   Sun May 5 16:59:47 2024 +0800

    ddd

commit b069423fe6bbe2d55ca6199cb4958740cb3e9f21
Author: Simon Yong <51854826+simony666@users.noreply.github.com>
Date:   Sun May 5 16:50:34 2024 +0800

    Squashed commit of the following:

    commit c66c7c4203b3189f57b07d55973425f367572368
    Author: Simon Yong <51854826+simony666@users.noreply.github.com>
    Date:   Sun May 5 14:22:04 2024 +0800

        Update .gitignore

    commit 5564c936a146fa2a69ebef71b38445036b83ccd8
    Author: Simon Yong <51854826+simony666@users.noreply.github.com>
    Date:   Sun May 5 13:56:43 2024 +0800

        Update .gitignore

    commit 7d13d63d05504d7555f62b22ac35b1108b86db8f
    Author: Simon Yong <51854826+simony666@users.noreply.github.com>
    Date:   Sun May 5 11:51:13 2024 +0800

        Update .gitignore

    commit ed68ec99f1ec3c78d3317ab2c6039f53e3b5e090
    Author: Simon Yong <51854826+simony666@users.noreply.github.com>
    Date:   Sun May 5 11:48:38 2024 +0800

        Delete Useless Item

    commit 212861e2d8a69751e2646295de1f54e8f1f32438
    Author: Simon Yong <51854826+simony666@users.noreply.github.com>
    Date:   Sun May 5 11:47:18 2024 +0800

        Update .gitignore

    commit bd4f424931c15e81c5d887e97ac8f84f03a48fe2
    Author: Zzz-anonymous <firsttry2468@gmail.com>
    Date:   Sun May 5 11:45:02 2024 +0800

        change path in ViewStart

    commit a4f85b06b8c06faca84cb76306bc0f8eee83a8cd
    Author: Zzz-anonymous <firsttry2468@gmail.com>
    Date:   Sun May 5 11:38:49 2024 +0800

        modify css, js, stafflayout

    commit 0e8ce94e558467bd0a37c93da82febe7ee6a705d
    Author: Zzz-anonymous <firsttry2468@gmail.com>
    Date:   Sun May 5 11:37:27 2024 +0800

        modify css, js, stafflayout

    commit 592a8d9b7010a4d81dfaf203829b1fe927bf9b86
    Author: Zzz-anonymous <firsttry2468@gmail.com>
    Date:   Sun May 5 09:52:29 2024 +0800

        add tutors, students, classes, subjects table

+++ Demo/Controllers/SubjectController.cs
@@ -23,6 +23,7 @@ namespace Demo.Controllers
         public ActionResult Index()
         {
             ViewBag.Subjects = db.Subjects;
+            ViewBag.Tutors = db.Tutors.ToDictionary(t => t.Id, t => t.Name);
             return View();
         }
 
@@ -31,7 +32,7 @@ namespace Demo.Controllers
         {
             //dropdown list for room type
             ViewBag.TutorList = new SelectList(db.Tutors.OrderBy(t => t.Id), "Id", "Name");
-            ViewBag.ClassesList = new SelectList(db.Classes.OrderBy(c => c.Id), "Id", "Name");
+            //ViewBag.ClassList = new SelectList(db.Classes.OrderBy(c => c.Id), "Id", "Name");
             return View();
         }
 
@@ -42,17 +43,14 @@ namespace Demo.Controllers
             {
                 vm.Id = NextId();
                 // Calculate end time based on start time and duration
-                vm.EndTime = vm.StartTime.AddHours(vm.Duration);
+                //vm.EndTime = vm.StartTime.AddHours(vm.Duration);
 
                 db.Subjects.Add(new()
                 {
                     Id = vm.Id,
                     Name = vm.Name,
                     Fees = vm.Fees,
-                    DayOfWeek = vm.DayOfWeek,
-                    StartTime = vm.StartTime,
-                    Duration = vm.Duration,
-                    EndTime = vm.EndTime,
+                    TutorId = vm.TutorId
                 });
                 db.SaveChanges();
 
@@ -67,10 +65,71 @@ namespace Demo.Controllers
         // Manually generate next id
         private string NextId()
         {
-            string max = db.Subjects.Max(s => s.Id) ?? "S000";
-            int n = int.Parse(max[1..]);
-            return (n + 1).ToString("'SU'000");
+            string max = db.Subjects.Max(s => s.Id);
+            if (max == null)
+            {
+                return "SU000"; // Return default value if no subjects exist yet
+            }
+            else
+            {
+                int n = int.Parse(max.Substring(2)); // Parse numeric part of the ID
+                return "SU" + (n + 1).ToString("000"); // Increment and format the ID
+            }
         }
+
+        // GET: Subject/Update
+        //[Authorize(Roles = "Member")]
+        public IActionResult Update(string id)
+        {
+            var s = db.Subjects.Find(id);
+            if (s != null)
+            {
+                ViewBag.TutorList = new SelectList(db.Tutors.OrderBy(t => t.Id), "Id", "Name");
+                var vm = new UpdateSubjectsVM
+                {
+                    Id = s.Id,
+                    Name = s.Name,
+                    Fees = s.Fees,
+                    TutorId = s.TutorId
+                };
+                return View(vm);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+        }
+
+        //[Authorize(Roles = "Member")]
+        [HttpPost]
+        public IActionResult Update(UpdateSubjectsVM vm)
+        {
+            if (ModelState.IsValid)
+            {
+                var s = db.Subjects.Find(vm.Id);
+
+                if (s != null)
+                {
+                    s.Name = vm.Name;
+                    s.Fees = vm.Fees;
+                    s.TutorId = vm.TutorId;
+                    db.SaveChanges();
+                    TempData["Info"] = "Subject updated.";
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    ModelState.AddModelError("Id", "Subject not found.");
+                }
+            }
+            else // If model state is invalid for POST request, reload the GET version of the page
+            {
+                return RedirectToAction("Update", new { id = vm.Id });
+            }
+
+            return View(vm);
+        }
+
 
         // POST: Subject/Delete
         [HttpPost]
@@ -80,9 +139,19 @@ namespace Demo.Controllers
 
             if (s != null)
             {
-                db.Subjects.Remove(s);
-                db.SaveChanges();
-                TempData["Info"] = $"Subject {s.Id} deleted.";
+                // Check if any students are associated with this class
+                var studentsWithSubjects = db.ClassesSubjects.Any(s => s.SubjectsId.Contains(id));
+
+                if (!studentsWithSubjects)
+                {
+                    db.Subjects.Remove(s);
+                    db.SaveChanges();
+                    TempData["Info"] = $"Class {s.Id} deleted.";
+                }
+                else
+                {
+                    TempData["Info"] = $"Cannot delete subject {s.Id} because there are students associated with it.";
+                }
             }
 
             return RedirectToAction("Index");
